enable clang as a compiler for cpu on any platform split up cpu/gpu exeuctioners to avoid limits fix sanitizers with clang to enable memory sanitizer standardize types

+++ libnd4j/include/system/common.h
@@ -176,7 +176,7 @@ namespace sd {
 
 using Pointer = void*;
 using LongType = long long;
-using UnsignedLong = uint64_t;
+using UnsignedLong = unsigned long long;
 using Unsigned = unsigned int;
